+++ src/transaction/index.ts
@@ -3,6 +3,7 @@
 // that can be found in the LICENSE.md file. All rights reserved.
 
 import * as index from '@edge/index-utils'
+import * as xe from '@edge/xe-utils'
 import { Network } from '../main'
 import { askSecure } from '../input'
 
@@ -19,6 +20,21 @@ export const askToSignTx = async (opts: { passphrase?: string }): Promise<void> 
     if (passphrase.length === 0) throw new Error('passphrase required')
     opts.passphrase = passphrase
     console.log()
+  }
+}
+
+export const handleCreateTxResult = (network: Network, result: xe.tx.CreateResponse): boolean => {
+  if (result.metadata.accepted !== 1) {
+    console.log('There was a problem creating your transaction. The response from the blockchain is shown below:')
+    console.log()
+    console.log(JSON.stringify(result, undefined, 2))
+    return false
+  }
+  else {
+    console.log('Your transaction has been submitted and will appear in the explorer shortly.')
+    console.log()
+    console.log(`${network.explorer.baseURL}/transaction/${result.results[0].hash}`)
+    return true
   }
 }
